scripts/requirements-images.txt
# Optional: used by scripts/compose-social-preview.py (GitHub social preview)
pillow>=10.0.0
